# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auction API",
        "description": "API for the Antique Items Auction Web Application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/bids/current": {
            "get": {
                "tags": [
                    "User Bids"
                ],
                "summary": "Get current user's active bids",
                "description": "Retrieves a list of active bids for the authenticated user",
                "operationId": "get_user_current_bids",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Bid3"
                                            }
                                        },
                                        "totalItems": {
                                            "type": "integer"
                                        },
                                        "itemsPerPage": {
                                            "type": "integer"
                                        },
                                        "totalPages": {
                                            "type": "integer"
                                        },
                                        "currentPage": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/user/bids/awarded": {
            "get": {
                "tags": [
                    "User Bids"
                ],
                "summary": "Get user's awarded items",
                "description": "Retrieves a list of items awarded to the authenticated user",
                "operationId": "get_user_awarded_items",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Bid3"
                                            }
                                        },
                                        "totalItems": {
                                            "type": "integer"
                                        },
                                        "itemsPerPage": {
                                            "type": "integer"
                                        },
                                        "totalPages": {
                                            "type": "integer"
                                        },
                                        "currentPage": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Bid3": {
                "required": [
                    "bidder",
                    "item",
                    "isAutoBid",
                    "amount",
                    "id",
                    "uuid",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "bidder": {
                        "$ref": "#/components/schemas/User6"
                    },
                    "item": {
                        "$ref": "#/components/schemas/Item7"
                    },
                    "bidTime": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "isAutoBid": {
                        "type": "boolean"
                    },
                    "amount": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            },
...
            "User6": {
                "required": [
                    "username",
                    "name",
                    "id",
                    "uuid",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "maxLength": 180,
                        "minLength": 3
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "default": []
                    },
                    "name": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            },
            "Item7": {
                "required": [
                    "name",
                    "medias",
                    "id",
                    "uuid",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "startingPrice": {
                        "type": "string",
                        "nullable": true
                    },
                    "auctionStartTime": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "auctionEndTime": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "medias": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ItemMedia5"
                        }
                    },
                    "id": {
                        "type": "integer"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            },
...
            "ItemMedia5": {
                "required": [
                    "id",
                    "uuid",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [],
...
}